requirements.txt
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.15.4
fsspec==2024.6.0
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/gitpython_1699254434447/work
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
jiter==0.4.2
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
setuptools==69.5.1
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.43.0
